Try to fix Markov Chain again
